[libc][docs] Rename nl_types macro header to match nl_types.h
