+++ libc/include/llvm-libc-macros/CMakeLists.txt
@@ -371,7 +371,7 @@ add_macro_header(
 add_macro_header(
   nl_types_macros
   HDR
-    nl-types-macros.h
+    nl_types-macros.h
 )
 
 add_macro_header(
