+++ libc/include/nl_types.yaml
@@ -3,9 +3,9 @@ standards:
   - posix
 macros:
   - macro_name: NL_SETD
-    macro_header: nl-types-macros.h
+    macro_header: nl_types-macros.h
   - macro_name: NL_CAT_LOCALE
-    macro_header: nl-types-macros.h
+    macro_header: nl_types-macros.h
 types:
   - type_name: nl_catd
 enums: []
